Change the generate of the FIPS HMAC to be after the dh_strip phase.
Before we tried to generate the .hmac file at the dh_install phase. This
would use the build version of the shared library of gnutls. However,
debian still changes the library and strips symbols out of it. This
leads the issue that the generate hmac isn't valided.

By introducing the +execute_after_dh_strip step we can generate a vaild
hmac that will be usable on the target system.