Fix/hmac generation (#2)
In this commit, we change the generation of the FIPS HMAC to be after the `dh_strip` phase.

Before, we tried to generate the .hmac file at the `dh_install` phase. This would use the build version of the shared library of GnuTLS. However, the build process still changes the library and strips symbols out of it. This leads to the issue that the generated HMAC isn't validated. By introducing the `execute_after_dh_strip` step, we can generate a valid HMAC that will be usable on the target system.

Relates: https://github.com/gardenlinux/gardenlinux/issues/3436